Updated docs for default build device
I changed the build description from hammerhead to bullhead as Google does not provide hammerhead binaries for android-8.1.0_r52 which is the version maruos 0.6 is based on.

Hence Bullhead is a more appropriate example.
I also updated the link as to where to find the android version in the manifest as it was not valid anymore

Signed-off-by: Lolo <elpoisot@hotmail.fr>

Added space

Signed-off-by: Lolo <elpoisot@hotmail.fr>